Refactor: new layouy and code clean-up
