--- hello.txt
@@ -1,3 +1,0 @@
-hello world
-
-This is global nexus
